pyproject.toml
[project]
name = "mcp-sharepoint"
version = "0.1.7"
description = "MCP Server to integrate with SharePoint"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "mcp>=1.2.1",
 "office365-rest-python-client>=2.6.1",
 "python-dotenv>=1.0.0",
 "pymupdf>=1.23.0",
 "pandas>=2.0.0",
 "openpyxl>=3.1.0",
 "python-docx>=1.1.0",
]

[[project.authors]]
name = "sofias tech"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"